requirements.txt
numpy
opencv-python
pyttsx3
dlib
scipy
pydantic
langchain-openai
langchain-core
langchain
langchain-community
openai
speechrecognition
pygame
pyaudio
python-dotenv
requests
duckduckgo_search
wikipedia
stt_tts
